fix(applyMask): Apply mask was not considering precision
If I had configured precision to be 2 digits, for example, when the input was empty and I typed the char "1", it was masking the value to "0.1" instead of "0.01", and the same logic to 3 or more digits of precision. Added verification to validate if the rawValue is zero and then, repeat the "0" char according to configured precision.