feat(applyMask): Adds support for min and max
This change adds min and max options to the config. I made them optional in the CurrencyMaskConfig interface to avoid making a breaking change. I added a max value to the demo.

The changes support a fully positive range (eg. 10 to 100) and a fully negative range (eg. -10 to -100). I verified that typing and pasting work correctly in both cases. It also gracefully handles an invalid range where min > max (uses max for both).

I added test cases and manually tested with multiple combinations.