feat(addNumber): Adds support for natural input mode
Adds support for input modes. The default input mode is FINANCIAL, which
works the same as it does today. The new input mode is NATURAL, which
behaves more like a text editor/spreadsheet. Numbers start to the left
of the decimal (aka the "dollar" amount). Typing a number to the right
of the decimal replaces the next number. For example, typing "123.4" results in "123.40" instead of "12.34".