fix: fix android chrome support
This change fixes some issues in Android Chrome. Android Chrome doesn't send keypress event, so we rely on input events instead. The old input event handler code was basically duplicating code in the keypress handler, but it wasn't as robust, especially after we added support for natural mode. For example it assumed the cursor was always at the end of the input, so it isn't possible to add or delete a digit in the middle of the number. It doesn't allow you to add digits at all in natural mode.

The new code delegates to the existing keypress handler when adding
digits, resulting in consistent functionality with desktop browsers. The
algorithm is simple. Figure out what character was added, restore the
old string, then call handleKeyPress with the new character.

I also tweaked the remove logic by setting the selection index before
calling remove number, so it's now possible to remove a digit at any
index.

In addition to adding unit tests, I tested this on an Android Simulator
and my personal android device

Addresses issues #90 and #100