fix(InputService): Enforce min and max when clearing value
Currently, clearing the input field of all numbers (eg. by backspacing the only digit)
while nullable==false will coerce the value to 0, even if you specified a min/max range
that doesn't include 0.

For example, if nullable = false, min = 5, max = 10, then deleting the
only digit will result in the value 0 instead of 5.  This only seems to happen when precision
is set to 0 (aka no decimal). This changes coerces the value to be within the min/max range.

Current Behavior when clearing the input:

|   | nullable == true | nullable == false |
| - | - | - |
| min/max not specified | null | always 0 (correct) |
| min or max specified  | null | always 0 (wrong) |

New Bahvior:

| | nullable == true | nullable == false |
| - | - | - |
| min/max not specified | null | always 0 (correct) |
| min or max specified | null | 0 if in range, otherwise the closest number in range (usually min, but could be max if the range is entirely negative) |