Improve changeToNegative in NATURAL mode
This change improves the typing of negative numbers in NATURAL mode when starting to type in an empty/zero-valued field.
In NATURAL mode the user expects that typing [-] [1] [2] [3] leads to a field value of  e.g. "$ -123".
With this change input handling is still not perfect in NATURAL mode, but it allows you to continue typing numbers after hitting [-] on an empty/zeroish input.